fix(backend): add delete modal

+++ packages/backend/src/_migrations/1759327920930-auto-migration.ts
@@ -1,26 +1,23 @@
 import { MigrationInterface, QueryRunner } from "typeorm";
+import { domifaConfig } from "../config";
 
 export class AutoMigration1759327920930 implements MigrationInterface {
   name = "AutoMigration1759327920930";
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "structure" DROP COLUMN "verified"`);
-    await queryRunner.query(
-      `ALTER TABLE "structure" ADD "statut" text NOT NULL DEFAULT 'EN_ATTENTE'`
-    );
-    await queryRunner.query(`ALTER TABLE "structure" ADD "statutDetail" text`);
-    await queryRunner.query(
-      `ALTER TABLE "structure" ALTER COLUMN "telephone" DROP DEFAULT`
-    );
-    await queryRunner.query(
-      `ALTER TABLE "structure" ALTER COLUMN "sms" SET DEFAULT '{"senderName": null, "senderDetails": null, "enabledByDomifa": true, "enabledByStructure": false, "schedule" :{ "monday": false "tuesday": true "wednesday": false "thursday": true "friday": false } }'`
-    );
-    await queryRunner.query(
-      `ALTER TABLE "user_structure" DROP COLUMN "fonctionDetail"`
-    );
-    await queryRunner.query(
-      `ALTER TABLE "user_structure" ADD "fonctionDetail" text`
-    );
+    if (
+      domifaConfig().envId === "prod" ||
+      domifaConfig().envId === "preprod" ||
+      domifaConfig().envId === "local"
+    ) {
+      await queryRunner.query(`ALTER TABLE "structure" DROP COLUMN "verified"`);
+      await queryRunner.query(
+        `ALTER TABLE "structure" ADD "statut" text NOT NULL DEFAULT 'EN_ATTENTE'`
+      );
+      await queryRunner.query(
+        `ALTER TABLE "structure" ADD "decision" jsonb NOT NULL DEFAULT '{}'::jsonb`
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
